+++ package.json
@@ -20,15 +20,17 @@
     "@mui/icons-material": "latest",
     "@mui/lab": "^5.0.0-alpha.114",
     "@mui/material": "latest",
-    "@next/font": "latest",
     "@mui/x-date-pickers": "^5.0.6",
+    "@next/font": "latest",
     "@types/js-cookie": "^3.0.3",
+    "@types/lodash": "^4.14.195",
     "@types/node": "18.14.6",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
     "js-cookie": "^3.0.5",
+    "lodash": "^4.17.21",
     "next": "13.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
